Fixing compile/linker issues with Tpetra

+++ src/vectors/trilinos/tpetra/TpetraVectorOperations.cpp
@@ -2,7 +2,7 @@
 
 namespace AMP::LinearAlgebra {
 
-template class TpetraVectorOperations<double, int32_t, long long>;
-template class TpetraVectorOperations<float, int32_t, long long>;
+template class TpetraVectorOperations<double, int32_t, int64_t>;
+template class TpetraVectorOperations<float, int32_t, int64_t>;
 
 } // namespace AMP::LinearAlgebra
